refactor(project): make injectable component an actual HOC
refactor(project): make injectable component example complete

refactor(project): remove injectable component example

refactor(project): make injectable component an actual HOC
