+++ packages/ui-react/src/components/Tag/Tag.tsx
@@ -1,13 +1,16 @@
 import React from 'react';
 import classNames from 'classnames';
 
-import type { TagProps } from '../../../types/components';
-import { TAG_COMPONENT } from '../../modules/types';
-import InjectableComponent from '../../modules/InjectableComponent';
-
 import styles from './Tag.module.scss';
 
-const DefaultTag: React.FC<TagProps> = ({ children, className, isLive = false, size = 'normal' }: TagProps) => {
+type TagProps = {
+  className?: string;
+  isLive?: boolean;
+  children?: React.ReactNode;
+  size?: 'normal' | 'large';
+};
+
+const Tag: React.FC<TagProps> = ({ children, className, isLive = false, size = 'normal' }: TagProps) => {
   return (
     <div
       className={classNames(className, styles.tag, styles[size], {
@@ -18,7 +21,5 @@ const DefaultTag: React.FC<TagProps> = ({ children, className, isLive = false, s
     </div>
   );
 };
-
-const Tag = (props: TagProps) => InjectableComponent(TAG_COMPONENT, props, DefaultTag);
 
 export default Tag;
