--- packages/ui-react/src/modules/types.ts
@@ -1,1 +1,0 @@
-export const TAG_COMPONENT = Symbol('TAG_COMPONENT');
