--- platforms/web/src/components/Tag/Tag.tsx
@@ -1,9 +1,0 @@
-const Tag = () => {
-  return (
-    <div>
-      <h1>Injected Tag</h1>
-    </div>
-  );
-};
-
-export default Tag;
